Add files via upload
K-Nearest Neighbor (kNN) Classifier is a supervised pattern classifier that determines the class of
an input sample based on the distance to k nearest labeled neighbors. kNNs are considered as a
type of lazy learning method where an evaluation is performed as needed. In this assignment, you
are going to implement basic kNN algorithm and analyze the effect of normalization.